+++ crates/common/src/lib.rs
@@ -14,6 +14,7 @@
 extern crate core;
 
 pub mod builtins;
+pub mod config;
 pub mod matching;
 pub mod model;
 pub mod tasks;